requirements.txt
python-telegram-bot==20.6
langchain-openai==0.3.16
